Update README.md
Added additional comments on which myhdl version to use, the first minor release 0.1.0 will target myhdl 1.0 but the master branch currently lags the myhdl 1.0 development (myhdl master branch).  This is mainly because of the addition of the `@myhdl.block` decorator that is required for each ~~module~~ block.  The rhea master branch will be updated once the API is more stable.