migrated to myhdl 1.0
In myhdl 1.0 a new decorated was added for modules - now
known as **blocks**.  The decorator is used on all functions
that are intended to describe digital circuits and be converted.
The addition of the block decorator changed the beharior, instead
of lists of generators (list of list of gens ...) a block now
returns a `Block` object.  The return of a myhdl.block function
is considered an instance.

In addition to the myhdl 1.0 migration some minor code clean-up
was completed and updates to documentation.