A VGA timing fix and additional vga tests and examples. (#36)
This pull-request includes a fix for the VGA core. For line rate was too low for higher resolutions. This sparked a change (that still needs to be completed) to use a VGATimingParameter class to define the timing and check the timing.

This pull-request also includes addition board tests, the VGA was tested on a zybo development board.

* changed default resolution to 800x600

* change the full vga screen to be an even number of lines

* updated video documentation

* don't create memory in the interface

* clean up timing parameters

* increased the line rate for the higher resolutions

* adjust front and back porch

* moving to VGATimingParameter object